fix : View 폴더 정리
