+++ src/features/auth/login/index.ts
@@ -1,2 +1,1 @@
 export { LoginForm } from './ui/login-form'
-export { NewLoginForm } from './ui/new-login-form'
