+++ src/features/auth/login/model/types.ts
@@ -1,3 +1,8 @@
+export type LoginPayload = {
+  email: string
+  password: string
+}
+
 export type LoginResponse = {
   id: number
   username: string
