+++ src/views/(loggedIn)/board/board-page.tsx
