+++ src/views/(loggedIn)/board/index.ts
