+++ src/views/(loggedIn)/board/mobile-board-page.tsx
