+++ src/views/(loggedIn)/home/mobile-main-page.tsx
