+++ src/views/(loggedOut)/email-validate/EmailValidatePage.tsx
