+++ src/views/(loggedOut)/email-validate/index.ts
