+++ src/views/(loggedOut)/email-validate/mobile-email-validate-page.tsx
