+++ src/views/(loggedOut)/invite-lab/index.ts
