+++ src/views/(loggedOut)/login/index.ts
@@ -1,2 +1,2 @@
-export { LoginPage } from './LoginPage'
 export { MobileLoginPage } from './mobile-login-page'
+export { LoginPage } from './login-page'
