+++ src/views/(loggedOut)/login/login-page.tsx
@@ -1,12 +1,10 @@
-'use client'
-
-import { NewLoginForm } from '@/features/auth/login'
+import { LoginForm } from '@/features/auth/login'
 import { AuthLinkItem } from '@/widgets/auth'
 
-const MyLabLoginPage = () => {
+export const LoginPage = () => {
   return (
     <>
-      <NewLoginForm />
+      <LoginForm />
       <footer className={'mt-3'}>
         <AuthLinkItem title={'회원가입하러 가기'} text={'MyLab이 처음이신가요?'} href={'/signup'} />
         <AuthLinkItem title={'비밀번호 재설정하기'} text={'비밀번호를 잊으셨나요?'} href={'#'} />
@@ -14,5 +12,3 @@ const MyLabLoginPage = () => {
     </>
   )
 }
-
-export default MyLabLoginPage
