+++ src/views/(loggedOut)/signup/index.ts
