+++ src/views/(loggedOut)/signup/mobile-signup-page.tsx
