+++ src/views/(loggedOut)/survey/SurveyPage.tsx
