+++ src/views/(loggedOut)/survey/index.ts
