+++ src/views/(loggedOut)/survey/mobile-survey-page.tsx
